+++ src/cpm/mod.rs
@@ -45,6 +45,7 @@ pub mod modem_port;
 pub mod tarbell;
 pub mod uart;
 pub mod wd1771;
+pub mod z80pack;
 
 pub use fcb::{parse_afn, parse_command_fcb, parse_dir_operand, split_8_3, Fcb, FCB_SIZE};
 pub use fs::{CpmFs, DEFAULT_DMA, NUM_DRIVES};